requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
certifi==2020.12.5
cffi==1.17.1
chardet==4.0.0
click==8.1.7
cryptography==43.0.1
Cython==3.0.11
distlib==0.3.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.114.0
fastapi-cli==0.0.5
filelock==3.16.0
flake8==7.1.1
future==0.18.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.4
jose==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
numpy==2.1.0
openpyxl==3.1.5
pandas==2.2.2
passlib==1.7.4
pep8==1.7.1
platformdirs==4.3.2
ply==3.11
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pyasn1==0.6.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
pyzk==0.9
requests==2.25.1
rich==13.8.0
rsa==4.9
schedule==1.2.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.4
starlette_pydantic==0.11
thriftpy2==0.5.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.3
uvicorn==0.30.6
uvloop==0.20.0
virtualenv==20.26.4
watchfiles==0.24.0
websockets==13.0.1
zk==0.8.5
